Node installation instructions are missing a step
It did not suffice for me to just add the node_modules/jasmine-expect/index.js line to the helpers array. I also had to add `require("jasmine-expect")` to my spec helper file. I just recently found this project and am setting it up for the first time. I don't know how it works under-the-hood, but I think this is useful info to have in the README in case others encounter the same situation.